requirements.txt
ag2==0.9.6
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asyncer==0.0.8
attrs==25.3.0
Authlib==1.6.0
banks==2.1.3
beautifulsoup4==4.13.4
Bottleneck==1.5.0
cachetools==5.5.2
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
cryptography==45.0.5
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
fastapi==0.116.1
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.5.1
google-adk==1.6.1
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.103.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.25.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
greenlet==3.2.3
griffe==1.7.3
grpc-google-iam-v1==0.14.2
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langchain==0.3.26
langchain-core==0.3.68
langchain-google-vertexai==2.0.27
langchain-text-splitters==0.3.8
langgraph==0.3.34
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.72
langsmith==0.4.5
llama-cloud==0.1.32
llama-cloud-services==0.6.43
llama-index==0.12.48
llama-index-agent-openai==0.4.12
llama-index-cli==0.4.4
llama-index-core==0.12.48
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.7.10
llama-index-instrumentation==0.2.0
llama-index-llms-google-genai==0.2.4
llama-index-llms-openai==0.4.7
llama-index-multi-modal-llms-openai==0.5.3
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.11
llama-index-readers-llama-parse==0.4.0
llama-index-workflows==1.1.0
llama-parse==0.6.43
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.11.0
multidict==6.6.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numexpr==2.11.0
numpy==2.3.1
openai==1.95.1
openinference-instrumentation==0.1.34
openinference-instrumentation-autogen==0.1.9
openinference-instrumentation-langchain==0.1.46
openinference-instrumentation-llama-index==3.3.3
openinference-semantic-conventions==0.1.21
opentelemetry-api==1.35.0
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-instrumentation==0.56b0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.2.3
pillow==11.3.0
platformdirs==4.3.8
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyparsing==3.2.3
pypdf==5.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.26.0
rsa==4.9.1
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sse-starlette==2.4.1
starlette==0.47.1
striprtf==0.0.26
tenacity==8.5.0
termcolor==3.1.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
validators==0.35.0
watchdog==6.0.0
websockets==15.0.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
